Move cells charts fetch to hook
